Update func_wrapper.py
Updated the implementation of `@handle_array_like` decorator as it was causing the Array API tests to fail for some reason. Using `inspect.signature(fn)` preserves the correct behavior of the decorator but was causing Array API tests to fail. On the other hand, using `typing.get_type_hints(fn)` does *not* preserve the correct behavior of the decorator but passes on the Array API tests. Hence, decided to to come up with a hackey solution and merge both these approaches together.